+++ necronet/CHANGELOG.md
@@ -2,6 +2,32 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [0.0.2](https://github.com/pabron7/necronet/compare/v0.0.1...v0.0.2) (2025-04-15)
+
+
+### Features
+
+* add ConsentScene ([d1a5dcc](https://github.com/pabron7/necronet/commit/d1a5dcc43020662959b61202b444406768ab6902))
+* add creator credits and custom events to footer ([6b204b2](https://github.com/pabron7/necronet/commit/6b204b24dd5ae68a3713b185c9203a988343807a))
+* add FetchedPeopleStore for managing people in local storage ([8dd298a](https://github.com/pabron7/necronet/commit/8dd298ac79de8fc3fb34e878fcaec4b8b6355a7a))
+* add firebase initialization ([efba1e7](https://github.com/pabron7/necronet/commit/efba1e7c3bbff9c02dc38969c9f85b20ff2a0f5e))
+* add LastActionStore to distribute the last action to the game ([5020f49](https://github.com/pabron7/necronet/commit/5020f49a02e447ff2fc50cb127a3ef5d2c6181a1))
+* add LocalStorageManager to centralize local storage management ([215398e](https://github.com/pabron7/necronet/commit/215398e915e94d60505fe64973c10d8669dd35ac))
+* add narrationUtils to populate narration sequence ([c5bf628](https://github.com/pabron7/necronet/commit/c5bf628d8b83c4caa2d6209cff5fdfaf14118b3d))
+* add peopleList to centralize fetched people listing ([ae29baf](https://github.com/pabron7/necronet/commit/ae29bafc9e802780c645af928c6b209d815cf84f))
+* add refresh link to the header ([b2a1562](https://github.com/pabron7/necronet/commit/b2a15622982ac84743cc56f15944beeab7a8c2fc))
+* add ResultsScene ([c672361](https://github.com/pabron7/necronet/commit/c672361b3554c16f10ef337d8674ebe05ba1e6d7))
+* add SelectedPersonStore to distribute selected person to the game ([574bd96](https://github.com/pabron7/necronet/commit/574bd96ab27e2c6d70ba3655fbc2b570a4b7896c))
+* add TaskSelection scene ([78b0f5e](https://github.com/pabron7/necronet/commit/78b0f5e749493f8cb709657cf06372e290913c61))
+* create person classes for type safety ([fd9ad4e](https://github.com/pabron7/necronet/commit/fd9ad4e9bdce0b6f8e3bcd155c7ef00299eaff6f))
+* update GameScene to integrate gameplay elements ([c931fc9](https://github.com/pabron7/necronet/commit/c931fc9488f1d2cb74d8e09ca6b11f3bf854a1bb))
+
+
+### Bug Fixes
+
+* import PersonalityType to PersonalityMapper ([564e8ac](https://github.com/pabron7/necronet/commit/564e8ac49ba6092c1a453e8c39a57bf621391051))
+* update resizer import location with correct path ([abc58d8](https://github.com/pabron7/necronet/commit/abc58d8c108d47ea0dc5c1315e03978e163befe6))
+
 ### 0.0.1 (2025-04-07)